requirements.txt
chalice==1.32.0
sqlalchemy>=1.4.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
pyjwt>=2.0.0
bcrypt>=4.0.0
email-validator>=2.0.0
flasgger>=0.9.7.1
